packages/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Shared Pydantic schemas for AI Lead Qualification Voice Agent System"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "pydantic>=2.0",
    "resend>=2.0",
    "firecrawl-py>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
scraping = [
    "firecrawl-py>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
